requirements.txt
Flask
telethon
python-dotenv